+++ src/index.ts
@@ -1,10 +1,10 @@
-import {registerPlugin} from "@capacitor/core";
+import { registerPlugin } from '@capacitor/core';
 
 import type { FileSharerPlugin } from './definitions';
 
 const FileSharer = registerPlugin<FileSharerPlugin>('FileSharer', {
-    web: () => import('./web').then(m => new m.FileSharerPluginWeb()),
-    // electron: () => ("./electron").then(m => new m.FileSharerPluginElectron())
+  web: () => import('./web').then(({ FileSharerPluginWeb }) => new FileSharerPluginWeb()),
 });
+
 export * from './definitions';
 export { FileSharer };