+++ src/web.test.ts
@@ -1,8 +1,0 @@
-describe('web plugin', () => {
-    it('convert to byteArray', async () => {
-        // WebUtils.toByteArray(oneDriveOptions).then(webOptions => {
-        //     expect(webOptions).not.toBeNull();
-        // });
-    });
-
-});